Working on GraphQL mutations
Before we can create a user and log them in we need to implement
the ability to create the PKI keys and perform all the necessary
oprations such as signing and encryption and decryption using
both private and public keys.

The /public-key api route has been created which returns the
server's public key.

A collection of methods have been implemented for RSA with  4096
bit modulus.
  - getPublicKey
  - getPrivateKey
  - generateRSAKeys
  - signWithPrivateKey
  - verifyWithPublicKey
  - encryptWithPublicKey
  - decryptWithPublicKey
  - decryptWithPrivateKey
  - encryptWithPrivateKey

All functionality has been implemented with TDD an all tests are
passing.

A create user mutation has been added, but there are no tests yet
as the remaining process has not yet been implemented. Since we
are using PKI, there are no passwords and a user will need to
upload their Public Key next.