Working on user creation.
A user can be created in the database. When a new account is
created, a signed token is returned with a nonce that can
be used to access the next step. This next step allows the user
to ensure they are communicating with the right server and to
upload their public key.