Implemented the account completeion phase. The front-end has not
been implemented to support this, but has been verifed through
TDD. Currently a new user can be created. Using the completion
invite, a user can upload their public key and an authSession
is returned.

Logic still needs to be implemented for session verification on
protected enpoints. A means for logging in, also needs to be
implemented.