Refactoring
Removed the individual scripts for running the test-scripts and
created a npm script called `ts` that runs the command necessary
to execute the .ts files from the command line. Now the test
scripts can be invoked using npm run ts ./path/to/tests arg1 arg2
Also corrected logic when checking the current number of user
passwords in the addPassword mutation.
Expanded the middleware to return more user account info in the
AuthContext- this is only visible server side